add tasklist update

+++ src/components/TaskItem.tsx
@@ -1,5 +1,6 @@
 import { useTasks } from '../context/useTasks';
 import { Task } from '../interfaces/task.interface';
+import { IoCheckmarkDone, IoTrash} from 'react-icons/io5';
 
 interface Props {
 	task: Task;
@@ -7,7 +8,7 @@ interface Props {
 
 function TaskItem({ task }: Props) {
 
-	const {deleteTask} = useTasks()
+	const {deleteTask, updateTask} = useTasks()
 
 	return (
 		<div key={task._id} className="bg-gray-900 p-2 my-2 flex 
@@ -17,10 +18,30 @@ function TaskItem({ task }: Props) {
 				<p>{task.description}</p>
 			</div>
 			<div className="flex gap-x-2">
-				<button>Update</button>
-				<button onClick={async() => {
+				{
+					task.done ? (
+						<IoCheckmarkDone 
+							className="text-green-500" 
+							onClick={() => {
+							updateTask(task._id, {
+								done: !task.done
+							});
+						}}
+						/>) : ( 
+						<IoCheckmarkDone 
+							className="text-gray-500" 
+							onClick={() => {
+							updateTask(task._id, {
+								done: !task.done
+							});
+						}}
+						/>
+					)
+				}
+				<IoTrash onClick={async() => {
+					if(!window.confirm("Are you sure you want to delete this task?")) return;
 					await deleteTask(task._id)
-				}}>Delete</button>
+				}}/>
 			</div>
 		</div>
 	);
